created a users/login endpoint. It requests a json body with username and password;
added a check to users/create to verify if a username already exists;

amended the users/getAll endpoint to only return a list of ids and usernames;

refactored the getAllUsers method in UserServiceImpl to include an empty check;

refactored TournamentController endpoints to return and require compact Jsons. When using Swagger, the sample Json bodies
provided by Swagger may contain MORE variables than actually needed for that particular endpoint. This is because the same
DTO is used for multiple endpoints. Check comments in Controllers to see what query bodies are required.

added a StageDTO for getting tournament stage AND match data and a StageAndMatchData endpoint in TournamentController;

added an addParticipants endpoint to TournamentController;

refactored project to not include a Participant Class, instead Match has two participant Strings. Refactored MatchService
and Controller accordingly, as well as other relevant classes and methods;